fix: split image-url.ts to separate server-only fs functions (#212)
Moves getHeroImage and getManifestHeroImage into image-url.server.ts
with 'server-only' import guard. Prevents Turbopack from bundling
Node.js fs module into client components, which was causing 500 errors
on all pages.
